refactor(engine): monotonic ULID + drop next_actions placeholder + split invoke.ts
I1 — ULID must be monotonic within same ms (ulid/spec §Monotonicity):
  - Track lastMs + lastRand at module scope; increment lastRand + 1n on
    same-ms call; draw fresh 10 randomBytes only when clock advances.
  - 80-bit overflow spins until Date.now() > lastMs (unreachable in
    practice at 10^24 events/ms, but spec-required).
  - Replaces weak 2-ms-gap test with a 2000-ID tight-loop strict-ascending
    assertion + fixed-clock same-ms sort test.

I2 — drop compile-time pre-baking of next_actions:
  - Remove `defaultNextActions` field from CompiledCommand entirely.
  - Remove instantiateNextActions placeholder round-trip.
  - execute() now calls defaultSuccessNextActions(site, cmd, {supportsPagination})
    once per invocation with the literal site/cmd. Adapters that legitimately
    carry the substring `${site}` are no longer silently rewritten, and
    paginated commands flow their --cursor hint through correctly.
  - AdapterCommand gains an optional `paginated?: boolean` to drive the
    pagination next-action.

M1 — split invoke.ts into src/engine/kernel/{types,ulid,compile,execute}.ts:
  - types.ts   (56 LOC)  — pure type surface, zero runtime deps
  - ulid.ts    (83 LOC)  — monotonic generator, self-contained
  - compile.ts (193 LOC) — ajv singleton, cache, compileAll, getCompiled
  - execute.ts (242 LOC) — buildInvocation + execute end-to-end
  - invoke.ts  (29 LOC)  — re-export shim for backward compat
